Last minute fixes to try and make the game in some way playable.

+++ Assets/Scripts/General/Managers/AudioManager.cs
@@ -104,6 +104,6 @@ public class AudioAsset
 {
     public string ClipId;
     public AudioClip Clip;
-    public int BeatsPerMinute;
+    public float BeatsPerMinute;
     public AudioSource AudioSource;
 }
